클래스를 나누어보았음
