+++ README.md
@@ -13,4 +13,4 @@
 - [x] 랜덤으로 0 - 9 출력하는 함수 구현
 - [x] 경기 구조 구현 - 전진하지 못해도 출력되도록 함
 - [x] 우승자 출력하는 함수 구현
-
+- [x] Racing 클래스를 만들어서 Car 기능에 직접적으로 속하지 않는 함수를 모두 옮겼다.
